# Conflicts:
#	.gitignore

+++ backend/src/main/java/at/qe/backend/services/UserxService.java
@@ -4,16 +4,20 @@ import at.qe.backend.configs.WebSecurityConfig;
 import at.qe.backend.exceptions.Userx.LastAdminException;
 import at.qe.backend.exceptions.Userx.UserAlreadyExistsException;
 import at.qe.backend.exceptions.Userx.UserDoesNotExistException;
+import at.qe.backend.exceptions.Userx.UserStillHasGreenhousesException;
+import at.qe.backend.models.AuditLog;
 import at.qe.backend.models.UserRole;
 import at.qe.backend.models.Userx;
 import at.qe.backend.repositories.UserxRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Collection;
 import java.util.Date;
@@ -28,10 +32,14 @@ import java.util.HashSet;
 @Service
 @Scope("application")
 public class UserxService {
-
+    @Autowired
+    private AuditLogService auditLogService;
     private final UserxRepository userRepository;
-    public UserxService(UserxRepository userRepository) {
+
+    @Autowired
+    public UserxService(UserxRepository userRepository, AuditLogService auditLogService) {
         this.userRepository = userRepository;
+        this.auditLogService = auditLogService;
     }
 
     /**
@@ -82,13 +90,22 @@ public class UserxService {
      */
     @PreAuthorize("hasAuthority('ADMIN')")
     public void deleteUser(Userx user) throws UserDoesNotExistException, LastAdminException {
+        AuditLog auditLog = auditLogService.createNewAudit("delete", "NA", "user", false);
         if (!userRepository.existsByUsername(user.getUsername())) {
             throw new UserDoesNotExistException();
         }
         if (user.getRoles().contains(UserRole.ADMIN) && userRepository.countUserxByRolesContaining(UserRole.ADMIN) <= 1) {
-            throw new LastAdminException();
+            throw new LastAdminException(HttpStatus.BAD_REQUEST, "Cannot delete last admin!");
+        }
+        if (!user.getGreenhouses().isEmpty()){
+            auditLog.setTargetID(user.getUsername());
+            auditLogService.saveAuditLog(auditLog);
+            throw new UserStillHasGreenhousesException(HttpStatus.BAD_REQUEST, "User is still responsible for at least one greenhouse.");
         }
         userRepository.delete(user);
+        auditLog.setSuccess(true);
+        auditLog.setTargetID(user.getUsername());
+        auditLogService.saveAuditLog(auditLog);
     }
 
 
@@ -99,6 +116,7 @@ public class UserxService {
 
     @PreAuthorize("hasAuthority('ADMIN') or principal.username eq #username")
     public Userx updateUser(String username, String firstname, String lastname, String email, Collection<UserRole> roles) throws UserAlreadyExistsException {
+        AuditLog auditLog = auditLogService.createNewAudit("update", username, "user", false);
         if (userRepository.existsByEmail(email) && !userRepository.findByEmail(email).getUsername().equals(username)) {
             throw new UserAlreadyExistsException();
         }
@@ -107,11 +125,14 @@ public class UserxService {
         user.setLastName(lastname);
         user.setEmail(email);
         user.setRoles(new HashSet<>(roles));
+        auditLog.setSuccess(true);
+        auditLogService.saveAuditLog(auditLog);
         return saveUser(user);
     }
 
     @PreAuthorize("hasAuthority('ADMIN')")
     public Userx createUser(String username, String firstname, String lastname, String email, Collection<UserRole> roles, String password) throws UserAlreadyExistsException {
+        AuditLog auditLog = auditLogService.createNewAudit("create", "NA", "user", false);
         if (userRepository.existsByUsername(username) || userRepository.existsByEmail(email)) {
             throw new UserAlreadyExistsException();
         }
@@ -122,6 +143,9 @@ public class UserxService {
         user.setEmail(email);
         user.setRoles(new HashSet<>(roles));
         user.setPassword(WebSecurityConfig.passwordEncoder().encode(password));
+        auditLog.setSuccess(true);
+        auditLog.setTargetID(username);
+        auditLogService.saveAuditLog(auditLog);
         return saveUser(user);
     }
 }
